chapter-4/remove-metadata/requirements.txt
ffmpeg-python
Pillow
pikepdf
tinytag
python-docx